requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
pydantic>=2.8.2
pydantic-settings>=2.3.4
httpx>=0.27.2
PyPDF2>=3.0.1
reportlab>=4.2.2
python-multipart>=0.0.9
pyahocorasick>=2.1.0